+++ Sources/PulseUI/Extensions/Pulse+Extensions.swift
@@ -65,6 +65,16 @@ extension NetworkTaskEntity {
             error: decodingError
         )
     }
+
+    /// - returns `nil` if the task is an unknown state. It may happen if the
+    /// task is pending, but it's from the previous app run.
+    func state(in store: LoggerStore) -> NetworkTaskEntity.State? {
+        let state = self.state
+        if state == .pending && self.session != store.session.id {
+            return nil
+        }
+        return state
+    }
 }
 
 extension LoggerMessageEntity {