+++ Sources/PulseUI/Features/Inspector/Cells/NetworkRequestStatusSectionView.swift
@@ -33,10 +33,10 @@ final class NetworkRequestStatusSectionViewModel {
     let requestViewModel: NetworkRequestInfoCellViewModel
     let errorDetailsViewModel: KeyValueSectionViewModel?
 
-    init(task: NetworkTaskEntity) {
-        self.status = NetworkRequestStatusCellModel(task: task)
+    init(task: NetworkTaskEntity, store: LoggerStore) {
+        self.status = NetworkRequestStatusCellModel(task: task, store: store)
         self.errorDescription = task.state == .failure ? task.errorDebugDescription : nil
-        self.requestViewModel = NetworkRequestInfoCellViewModel(task: task)
+        self.requestViewModel = NetworkRequestInfoCellViewModel(task: task, store: store)
         self.errorDetailsViewModel = KeyValueSectionViewModel.makeErrorDetails(for: task)
     }
 }
@@ -49,7 +49,7 @@ struct NetworkRequestStatusSectionView_Previews: PreviewProvider {
             List {
                 ForEach(MockTask.allEntities, id: \.objectID) { task in
                     Section {
-                        NetworkRequestStatusSectionView(viewModel: .init(task: task))
+                        NetworkRequestStatusSectionView(viewModel: .init(task: task, store: .mock))
                     }
                 }
             }